interface: déplacer histogramme() dans scrutin/graphiques.py
La couture données / présentation prévoit que les figures Plotly vivent
dans graphiques.py, et views.py reste minuscule. L'histogramme (et son
helper clean_name) déménagent tels quels, sans changement de
comportement ; views.py se réduit à home_view.

PLAN_MODERNISATION.md ne liste donc plus que le déménagement de la carte
vers carte/figure.py comme travail différé.
